+++ src/components/AreaController.tsx
@@ -1,0 +1,11 @@
+import KeyController from "../controllers/KeyController";
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const AreaController: React.FC<Props> = ({ children }) => (
+  <KeyController>{children}</KeyController>
+);
+
+export default AreaController;
